==> Code/Validation/0_validation_approach.Rmd <==
---
title: "Validation"
author: "Grace Smith-Vidaurre"
date: "2024-01-17"
output: html_document
---

Purpose: Plan out a validation based on manual video scoring of videos recorded by the ABISSMAL tracking system. We decided to select all videos recorded in each of 3 days that represented major behavioral epochs during offspring development, including when eggs were laid, when chicks hatched, and when chicks fledged (1 day per epoch). We also randomly selected a number of videos recorded on each day, to ensure that the videos chosen for validation represented movements recorded across all days of data collection. 

```{r libraries and paths}

rm(list = ls())

library(tidyverse)
library(pbapply)
library(data.table)

orig_path <- "/media/gsvidaurre/Anodorhynchus/Data_Testing/Box_03_23Jun2023"

```

Get the directories with videos of interest.
```{r}

# Print all of the directories with video files
vid_dirs <- list.files(file.path(orig_path, "Data/Video"), pattern = "^2023_*", full.names = TRUE)

# Filter out dates with videos recorded prior to the first detection of a PIT tag on June 28th 2023
vid_dirs <- vid_dirs[!grepl(paste(paste("_", c("06_23", "06_24", "06_25", "06_26", "06_27"), "$", sep = ""), collapse = "|"), vid_dirs)]

# 50 days of data collection remain
vid_dirs

```

How many videos were recorded for the pair in Box_03? The first date that a PIT tag was recorded by the RFID system was 6-28 (June 28th, see the script for making figures).
```{r}

all_videos <- list.files(vid_dirs, pattern = ".mp4")

# 81,298 videos recorded over 50 days
length(all_videos)

# 40649 unique video recording events (since each motion detection event resulted in recording 1 pre- and 1 post-motion detection video) 
length(all_videos)/2

```

How many videos were recorded per day? Here focus on the pre-trigger videos to get the unique number of video recording events.
```{r}

# i <- 1
vids_by_day <- data.table::rbindlist(pblapply(1:length(vid_dirs), function(i){
  
  vids <- list.files(vid_dirs[i], pattern = ".mp4", recursive = TRUE)
  
  dates <- sapply(1:length(vids), function(x){
    
    tmp <- strsplit(vids[x], split = "_")[[1]]
    
    return(paste(tmp[4], tmp[5], sep = "_"))
    
  }) 
  
  return(
    data.frame(
      # These videos were saved prior to fixing a bug that placed videos from the day before into a folder for the given day
      # So get the day from the video file name instead
      # day = gsub(file.path(orig_path, "Data/Video/2023_"), "", vid_dirs[i]),
      date = dates,
      videos = vids
    )
  )
  
}))

glimpse(vids_by_day) 

# The minimum number of recording events each day was 17, and the max was 2258. Median of 720
vids_by_day %>%
  dplyr::filter(grepl("pre_trigger", videos)) %>% 
  group_by(date) %>%
  dplyr::summarise(
    n = length(videos)
  ) %>% 
  dplyr::summarise(
    min = min(n),
    median = median(n),
    max = max(n)
  )

```

Create a new directory on the hard drive for validation / manual video scoring.
```{r}

dir.create("/media/gsvidaurre/Anodorhynchus/ABISSMAL_MethodsPaper/Validation")
dir.create("/media/gsvidaurre/Anodorhynchus/ABISSMAL_MethodsPaper/Validation/videos")
dir.create("/media/gsvidaurre/Anodorhynchus/ABISSMAL_MethodsPaper/Validation/masked_videos")

```

# Selecting videos by behavioral epoch

Behavioral epochs determined for the pair of birds used in the methods manuscript: eggs were laid from July 4th to July 8th (07-04 to 07-08), eggs hatched from July 19th to 22nd (07-19 to 07-22), and chicks fledged from August 6th to 7th (08-06 to 08-07). See the script for making figures for the methods manuscript for more information. 

Pick one day for each of these epochs for manual video scoring. All of the videos recorded per day will be scored for each of these days. I picked the last day of each epoch above.
```{r}

days <- c("7_8", "7_22", "8_7")

vids_by_day %>%
  dplyr::filter(grepl("pre_trigger", videos)) %>%
  dplyr::filter(date %in% days) %>% 
  group_by(date) %>%
  dplyr::summarise(
    n = length(videos)
  )

# # A tibble: 3 × 2
#   date      n
#   <chr> <int>
# 1 7_22    710
# 2 7_8     725
# 3 8_7    2140

```

There are 3575 unique recording events across these 3 days, for a total of at least 7150 videos (pre- and post-). Together, the videos per recording event are a total of 15 seconds, meaning that 53,625 seconds or 894 minutes of video must be scored. If it takes 5 minutes to score the pair of videos representing each video recording event, then we will need to spend 4,470 minutes or 74.5 hours scoring these videos (this is an over-estimate).
```{r}

# The total number of videos across these 3 days is 7150, as expected
vids_by_day %>%
  dplyr::filter(date %in% days) %>% 
  nrow(.)

# Select these video recordings
epoch_vids <- vids_by_day %>%
  # Select the pre-trigger videos to represent the unique recording events
  dplyr::filter(grepl("pre_trigger", videos)) %>%
  dplyr::filter(date %in% days)

nrow(epoch_vids)
  
```

# Selecting videos across all other days

We also decided to score videos randomly sampled across all days, such that the videos scored for system validation are also representative of the broader temporal trajectory of data collection. If we select 10 unique recording events per each of the remaining 47 days of data collection after selecting videos by behavioral epoch, then we need to score 470 unique recording events (940 videos total).
```{r}

n <- 10

set.seed(879)

full_period_vids <- vids_by_day %>%
  # Do not include the days selected by behavioral epoch
  dplyr::filter(!date %in% days) %>% 
  # Select the pre-trigger videos to randomly sample the unique recording events
  dplyr::filter(grepl("pre_trigger", videos)) %>%
  group_by(date) %>% 
  nest() %>% 
  ungroup() %>%
  dplyr::mutate(
    rsamp_vids = purrr::map2(data, n, sample_n, replace = FALSE)
  ) %>%
  dplyr::select(-data) %>%
  unnest(rsamp_vids)
  # Checking, looks good
  # group_by(date) %>%
  # dplyr::summarise(n = n()) %>%
  # pull(n) %>%
  # unique()

glimpse(full_period_vids)

```

Next, combine the pre-trigger videos selected for the behavioral epoch scoring and the full data collection period scoring. There are 4,045 unique recording events now.
```{r}

target_vids <- epoch_vids %>% 
  bind_rows(full_period_vids)

glimpse(target_vids)

```

Randomize the order of these videos by unique recording event.
```{r}

set.seed(879)
index <- sample(1:nrow(target_vids), nrow(target_vids), replace = FALSE)

# 4045, looks good
length(index)

target_vids_rs <- target_vids %>% 
  dplyr::mutate(
    index = index
  ) %>%
  dplyr::arrange(-desc(index))

glimpse(target_vids_rs)

```

Then mask the names of these videos, add back the post-movement trigger videos and mask those names too.
```{r}

target_vids_rsm <- target_vids_rs %>%
  dplyr::mutate(
    masked_name = paste(paste(paste("video", index, sep = "_"), "1", sep = "_"), ".mp4", sep = "")
  ) %>% 
  # Add back the post-trigger videos while ensuring that each post- video follows the corresponding pre-trigger video
  bind_rows(
    target_vids %>%
      dplyr::mutate(
        videos = gsub("pre_", "post_", videos)
      ) %>% 
      dplyr::mutate(
        index = index
      ) %>% 
      dplyr::arrange(-desc(index)) %>%
      dplyr::mutate(
        masked_name = paste(paste(paste("video", index, sep = "_"), "2", sep = "_"), ".mp4", sep = "")
      )
  ) %>% 
  dplyr::arrange(-desc(index)) %>% 
  rowid_to_column() %>% 
  dplyr::select(rowid, videos, masked_name)

# Looks great
glimpse(target_vids_rsm)
# View(target_vids_rsm)

# Looks good, 8090 total videos to score and 4045 unique recording events
nrow(target_vids_rsm) == (nrow(target_vids_rs)*2)

```

Write out a spreadsheet of all of the video file names that will be scored, plus the column of new file names for masking purposes. Note that the order of videos was randomized prior to masking. Masking metadata about each video and also presenting these out of consecutive order to an observer should help avoid some forms of bias. This spreadsheet will be written to a folder that the observer will not see.
```{r}

target_vids_rsm %>% 
  write.csv(file.path("/media/gsvidaurre/Anodorhynchus/ABISSMAL_MethodsPaper/Validation/unmasked_video_names.csv"), row.names = FALSE)

```

Next, copy the videos that will be scored to another folder on the same hard drive.
```{r}

# Get all file names and paths for all videos
all_videos <- list.files(vid_dirs, pattern = ".mp4", full.names = TRUE)
head(all_videos)

# Loop over all target videos to get the path per video from the full list of paths
full_paths <- unlist(pblapply(1:nrow(target_vids_rsm), function(i){
  
  return(all_videos[grep(paste(paste(target_vids_rsm$videos[i], "$", sep = ""), collapse = "|"), all_videos)])
  
}))

# Check that the paths match the correct files, looks good
# data.frame(full_paths, vids = target_vids_rsm$videos) %>% 
  # View()

# Copy each video over to the correct folder on the main external hard drive 
file.copy(full_paths, file.path("/media/gsvidaurre/Anodorhynchus/ABISSMAL_MethodsPaper/Validation/videos", target_vids_rsm$videos))

# Check that the files copied over are the same as the target videos
copied_files <- list.files("/media/gsvidaurre/Anodorhynchus/ABISSMAL_MethodsPaper/Validation/videos", pattern = ".mp4")

# Looks good
all(copied_files %in% target_vids_rsm$videos)
all(target_vids_rsm$videos %in% copied_files)

```

Then rename these videos to the masked names, while creating another copy of each file. 
```{r}

head(target_vids_rsm$videos)
head(target_vids_rsm$masked_name)

file.copy(
  from = file.path("/media/gsvidaurre/Anodorhynchus/ABISSMAL_MethodsPaper/Validation/videos", target_vids_rsm$videos), 
  to = file.path("/media/gsvidaurre/Anodorhynchus/ABISSMAL_MethodsPaper/Validation/masked_videos", target_vids_rsm$masked_name)
)

```

These masked files can now be copied to another hard drive for sharing with observer(s) who will score these videos.

Summer Eckhardt scored the masked videos in BORIS using an ethogram that the group of co-authors developed for scoring coarse-grained behavioral states (perching, entering, exiting, inside of the container, or no birds inside of the container) for different subjects (adult female, adult male, juveniles). More information about the ethogram and manual video scoring is stored in a standard operating protocol for BORIS video scoring for this project.
