==> Code/Validation/7_validation_analysis_individual_identity.Rmd <==
---
title: "Video validation analysis 4: Individual identity inference"
author: "Grace Smith-Vidaurre"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output: html_document
---

*Purpose of individual identity detection*: Next, ask how reliably we captured individual identity with RFID compared to the manually scored videos. In this analysis, we will rely on the RFID data for the automated detection of individuals, which will have data for the PIT-tagged adults only. Focus this analysis on asking how often each PIT tag was detected for each adult compared to the manual scoring, and then both PIT tags when manual scoring identified activity by both adults (three overall levels of individual identity detection performance).

```{r setup, include = FALSE}

knitr::opts_chunk$set(echo = TRUE, eval = TRUE, warning = FALSE, messsage = FALSE)

```

Load packages and set paths.
```{r package and paths}

# Clean the global environment
rm(list = ls())

X <- c("tidyverse", "pbapply", "data.table", "tidyquant")

# Install the packages in X if not already installed
is_installed <- function(p) is.element(p, installed.packages()[,1])

invisible(lapply(1:length(X), function(x){
  if(!is_installed(X[x])){
    install.packages(X[x], repos = "http://lib.stat.cmu.edu/R/CRAN")
  }
}))

invisible(lapply(X, library, character.only = TRUE))

# Overall path: contains the aggregated BORIS spreadsheet in wide format across all video recording events
path <- "~/Desktop/ABISSMAL_validation_data"

# Path to the combined raw movement sensor data
sensor_path <- file.path(path, "raw_combined")

# Initialize a folder where we will save the indexed raw data for the validation analyses
indexed_dir <- "indexed_labeled_data"
indexed_path <- file.path(path, indexed_dir)

# Create this folder if it doesn't already exist
if(!dir.exists(indexed_path)){
  dir.create(indexed_path)
}

# Initialize the GitHub repo path, which holds a .csv file of a subsample of video file names used for code development
git_path <- "~/Desktop/GitHub_repos/ABISSMAL-methods-paper/Code"

seed <- 888

```

Source functions for calculating traditional performance metrics (precision, sensitivity, and specificity).
```{r}

source(file.path(git_path, "Validation", "utilities.R"))

```

TKTK need manual_scoring_nm too, and potentially other objects

4. INDIVIDUAL IDENTITY DETECTION: Next, ask how reliably we captured individual identity with RFID compared to the manually scored videos. In this analysis, we will rely on the RFID data for the automated detection of individuals, which will have data for the PIT-tagged adults only. Focus this analysis on asking how often each PIT tag was detected for each adult compared to the manual scoring, and then both PIT tags when manual scoring identified activity by both adults (three overall levels of individual identity detection performance).

General approach to calculating individual identity detection performance:

**True positive** = The same individual(s) was detected by both ABISSMAL and manual scoring.

**False negative** = ABISSMAL detected the incorrect PIT tag compared to the individual identity detected by manual scoring. 

**False positive** = ABISSMAL detected the incorrect PIT tag compared to the individual identity detected by manual scoring.

**True negative** = Neither ABISSMAL nor manual scoring detected activity by an adult.

Overall, we can assess both the precision (out of the positive predictions, which are truly positive) and sensitivity (or recall, what percentage of the total positives were predicted) of activity location inferences by ABISSMAL, as well as the the specificity (true negative rate, or how well the true negatives are predicted while avoiding false alarms). Note that false negatives and positives will be the same because they are defined as a mismatch in location scores between methods.

Assess the precision, sensitivity, and specificity of individual identity detection via the RFID module used by ABISSMAL.

First we need to condense the automated scoring and the manual scoring data frames to have a single row per video while retaining information about unique adult identities (from PIT tags in the ABISSMAL data, and adult subject labels in the manually scored data).
```{r}

# View(manual_scoring_nm)
names(manual_scoring_nm)


########### Convert the subject labels per dataset

# The subject names for individually identified adults in the manually scored data are "Female", "Male".
# Subject names for no detections of adults or unidentifiable adults will be all other subject labels: "Juveniles_Pooled", "Juvenile", "No_Birds", "Bird_of_Unknown_Age" (to be conservative, since such subjects could have been juvenile), and "Unknown_Adult".

# The unique PIT tag IDs for adults in this dataset are: "01-10-3F-8F-F0" (Male) and "01-10-16-CD-1C" (Female). See the script for creating figures that was written previously.
# unique(scored_clusters$individual_initiated)

# Convert the subject labels in the manually scored dataset to 3 labels: Female, Male, or No Identifiable Adult
manual_scoring_nm2 <- manual_scoring_nm %>% 
  # Replace the subject labels as above
  dplyr::mutate(
    Subject_mod = ifelse(Subject %in% c("Juveniles_Pooled", "Juvenile", "No_Birds", "Bird_of_Unknown_Age", "Unknown_Adult"), "No Identifiable Adult", Subject)
  )

# Looks good, only 3 values remain as expected 
unique(manual_scoring_nm2$Subject_mod)


# Get one row of PIT tag IDs per video recording event
glimpse(auto_scoring2)

# Create a new column that condenses PIT tag IDs from inferences about which individual started a movement event, which individual ended the movement event, and which individual was detected in a perching event
auto_scoring3 <- auto_scoring2 %>% 
  dplyr::mutate(
    individual_initiated = gsub("01-10-3F-8F-F0", "Male", individual_initiated),
    individual_initiated = gsub("01-10-16-CD-1C", "Female", individual_initiated),
    individual_ended = gsub("01-10-3F-8F-F0", "Male", individual_ended),
    individual_ended = gsub("01-10-16-CD-1C", "Female", individual_ended),
    perching_PIT_tag = gsub("01-10-3F-8F-F0", "Male", perching_PIT_tag),
    perching_PIT_tag = gsub("01-10-16-CD-1C", "Female", perching_PIT_tag)
  ) %>% 
  dplyr::mutate(
    Adult_IDs = paste(individual_initiated, individual_ended, perching_PIT_tag, sep = " - ")
  )

glimpse(auto_scoring3)
unique(auto_scoring3$Adult_IDs)

# Convert the subject labels in the ABISSMAL scored dataset to binary labels: Adult or No Adult
auto_scoring4 <- auto_scoring3 %>% 
  # Replace the Adult_ID values to be similar to the manually scored data
  dplyr::mutate(
    Subject_mod = ifelse(grepl("NA - NA - NA", Adult_IDs), "No Identifiable Adult", Adult_IDs)
  )

# Looks good. These labels will include the same subject repeated or more than one subject
unique(auto_scoring4$Subject_mod)

########### Convert the manual and automated scoring data frames to tables

# Get one row of subject labels per video recording event
manual_table <- manual_scoring_nm2 %>% 
  group_by(unmasked_video_file) %>%
  # Create a unique row identifier within videos
  dplyr::mutate(
    row_id = row_number()
  ) %>% 
  ungroup() %>% 
  # View()
  dplyr::select(row_id, unmasked_video_file, Subject_mod) %>% 
  pivot_wider(names_from = unmasked_video_file, values_from = Subject_mod) %>% 
  t()


auto_table <- auto_scoring4 %>% 
  group_by(assigned_recording_event) %>%
  # Create a unique row identifier within videos
  dplyr::mutate(
    row_id = row_number()
  ) %>% 
  ungroup() %>% 
  # View()
  dplyr::select(row_id, assigned_recording_event, Subject_mod) %>% 
  pivot_wider(names_from = assigned_recording_event, values_from = Subject_mod) %>% 
  t()

# View(auto_table)
# View(manual_table)

```

In this analysis, we need to remove video recording events that only included activity scored inside of the container, for a fair validation of the RFID system. The RFID antennae is mounted in the container entrance, so it can only detect activity by individuals at this location. We included videos that had activity scored for birds of an unknown age, but did not consider this subject label an adult detection to be conservative.
```{r}

# Iterate over all the unique videos included in the automated and manual scoring tables to find true positives and false negatives
# Remove video recording events that had INSC scores only
# unique(manual_scoring_nm$Behavior)

unique_videos <- manual_scoring_nm %>%
  # This filter will drop any rows with the INSC label
  # But since there are multiple rows per unique recording event, video recording events that had other labels will be retained in the unique() call below
  dplyr::filter(!grepl("INSC", Behavior)) %>%
  pull(unmasked_video_file) %>% 
  unique()

unique_videos

glimpse(auto_table)
glimpse(manual_table)

# These labels correspond to the 3 levels of validation of individual identity detection: detecting of the adult female, the adult male, and both adults together
adults <- c("Female", "Male", "Both")
labels <- list(
  "Female" = "Female",
  "Male" = "Male",
  "Both" = c("Female", "Male")
)

labels

# testing
i <- 1
n <- 3

tp_fp_tn_fn <- data.table::rbindlist(pblapply(1:length(unique_videos), function(i){
  
  # Get unique behavioral scores for the given unique recording event from summarized tables of both scoring methods
  auto_scores <- unique(auto_table[grep(unique_videos[i], dimnames(auto_table)[[1]]), ])
  auto_scores <- auto_scores[!is.na(auto_scores)]
  
  manual_scores <- unique(manual_table[grep(unique_videos[i], dimnames(manual_table)[[1]]), ])
  manual_scores <- manual_scores[!is.na(manual_scores)]
  
  auto_scores
  manual_scores
  
  # Iterate over adult labels
  res <- data.table::rbindlist(lapply(1:length(adults), function(n){
    
    # Convert scores to Boolean using the labels associated with the respective location
    if(!adults[n] %in% "Both"){
      
      auto <- ifelse(any(grepl(labels[[adults[n]]], auto_scores)), TRUE, FALSE)
      manual <- ifelse(any(grepl(labels[[adults[n]]], manual_scores)), TRUE, FALSE)
      
      # auto
      # manual
      
      # If checking for labels in both locations, then loop over the location labels for a thorough Boolean conversion
    } else {
      
      tmp_labs <- labels[[adults[n]]]
      
      tmp_res <- data.table::rbindlist(lapply(1:length(tmp_labs), function(z){
        
        auto <- any(grepl(tmp_labs[z], auto_scores))
        manual <- any(grepl(tmp_labs[z], manual_scores))
        
        return(
          data.frame(
            auto = auto,
            manual = manual
          )
        )
        
      }))
      
      auto <- all(tmp_res$auto)
      manual <- all(tmp_res$manual)
      
    }
    
    # Score true positives
    if(auto & manual){
      
      true_positive <- 1
      
    } else {
      
      true_positive <- 0
      
    }
    
    # Score false positives
    if(auto & !manual){
      
      false_positive <- 1
      
    } else {
      
      false_positive <- 0
      
    }
    
    # Score true negatives
    if(!manual & !auto){
      
      true_negative <- 1
      
    } else {
      
      true_negative <- 0
      
    }
    
    # Score false negatives
    if(manual & !auto){
      
      false_negative <- 1
      
    } else {
      
      false_negative <- 0
      
    }
    
    return(
      data.frame(
        unique_video_recording_event = unique_videos[i],
        adult_labels = adults[n],
        true_positive = true_positive,
        false_positive = false_positive,
        true_negative = true_negative,
        false_negative = false_negative
      )
    )
    
  }))
  
  return(res)
  
}))

# There should be 3 rows per unique video recording event after the filtering above, looks good
length(unique_videos) * length(adults) == nrow(tp_fp_tn_fn)
glimpse(tp_fp_tn_fn)
# View(tp_fp_tn_fn)

```

Calculate precision, recall, sensitivity, and specificity of activity detection by adult.
```{r}

tp_fp_tn_fn_sum <- tp_fp_tn_fn %>% 
  group_by(adult_labels) %>% 
  dplyr::summarise(
    tps = sum(true_positive),
    fps = sum(false_positive),
    tns = sum(true_negative),
    fns = sum(false_negative)
  ) %>% 
  ungroup()

tp_fp_tn_fn_sum

# Precision across adults
precision_ind <- sapply(1:length(adults), function(i){
  
  precision(X = tp_fp_tn_fn_sum[grep(adults[i], tp_fp_tn_fn_sum$adult_labels), ], tps_col_nm = "tps", fps_col_nm = "fps")
  
})

# names(precision_ind) <- adults
precision_ind

# Sensitivity across adults
sensitivity_ind <- sapply(1:length(adults), function(i){
  
  sensitivity(X = tp_fp_tn_fn_sum[grep(adults[i], tp_fp_tn_fn_sum$adult_labels), ], tps_col_nm = "tps", fns_col_nm = "fns")
  
})

sensitivity_ind

# Specificity across adults
specificity_ind <- sapply(1:length(adults), function(i){
  
  specificity(X = tp_fp_tn_fn_sum[grep(adults[i], tp_fp_tn_fn_sum$adult_labels), ], tns_col_nm = "tps", fps_col_nm = "fns")
  
})

specificity_ind

# Note that sensitivity and specificity are the same, given the way that false positives and negatives were defined

data.frame(
  adult_labels = adults,
  precision = precision_ind,
  sensitivity = sensitivity_ind,
  specificity = specificity_ind
)

#   adult_labels precision sensitivity specificity
# 1       Female         0           0           0
# 2         Male       100          50          50
# 3         Both         0           0           0

```
